fix(sound): （自测）修复音量在115时，实际任务栏tips提示114%的问题
属于未发现问题，double类型转int精度丢失导致

Log: 任务栏音量tips提示与实际音量不符
Bug: https://pms.uniontech.com/zentao/bug-view-16696.html

+++ plugins/sound/sounditem.cpp
@@ -216,8 +216,9 @@ void SoundItem::refreshTips(const bool force)
         if (m_sinkInter->volume() * 1000 < m_applet->volumeValue())
             value = QString::number(m_applet->volumeValue() / 10) + '%';
         else
-            value = QString::number(int(m_sinkInter->volume() * 100)) + '%';
+            value = QString::number(m_sinkInter->volume() * 100) + '%';
     }
+
     m_tipsLabel->setText(QString(tr("Volume %1").arg(value)));
 }
